feat(bag-migration): step 5 — rename DerivaMLDatabase → DerivaMLBagView
The class previously named ``DerivaMLDatabase`` is more precisely a
*view* over a bag's contents than an alternative database
implementation. Per ADR-0006's three-class consumer layer:

    BagDatabase           (generic, upstream)
        ↑
    DatabaseModel         (deriva-ml-extended, step 4)
        ↑
    DerivaMLBagView       (domain-protocol-shaped — this commit)

Concrete changes:

src/deriva_ml/model/deriva_ml_bag_view.py: the canonical home for
the class, formerly src/deriva_ml/model/deriva_ml_database.py. The
class itself is renamed; the docstring captures the rename's
rationale.

src/deriva_ml/model/deriva_ml_database.py: replaced with a thin
re-export shim. ``DerivaMLDatabase`` is preserved as a legacy
alias for ``DerivaMLBagView`` so external callers keep working
for one release cycle.

model/__init__.py: the lazy ``__getattr__`` resolves both names
to the same class. ``__all__`` exposes both.

Internal callers (interfaces.py, dataset/dataset.py,
dataset/dataset_bag.py, execution/execution.py) renamed to
DerivaMLBagView. Tests (test_download.py, test_clone_catalog.py,
test_database.py) renamed for consistency.

No public API change — every import path that worked before still
works, either via the new canonical name or via the legacy alias.
